fix(semantic-search): move helper to ext
- moved label to l18n
- moved URL helper to ext

JIRA: CQ-582
risk: low
