+++ libs/sdk-ui-semantic-search/src/hooks/useSemanticSearch.ts
@@ -8,6 +8,7 @@ import {
 } from "@gooddata/sdk-model";
 import { useBackendStrict, useWorkspaceStrict } from "@gooddata/sdk-ui";
 import { IAnalyticalBackend } from "@gooddata/sdk-backend-spi";
+import { getUIPath } from "../utils/getUIPath.js";
 
 /**
  * The result of the semantic search hook.
@@ -171,31 +172,8 @@ const errorMessage = (e: unknown) => {
 const withUrl =
     (workspaceId: string) =>
     (item: ISemanticSearchResultItem): ISemanticSearchResultItemWithUrl => {
-        switch (item.type) {
-            case "dashboard":
-                return { ...item, url: `/dashboards/#/workspace/${workspaceId}/dashboard/${item.id}` };
-            case "visualization":
-                return { ...item, url: `/analyze/#/${workspaceId}/${item.id}/edit` };
-            case "metric":
-                return { ...item, url: `/metrics/#/${workspaceId}/metric/${item.id}` };
-            case "dataset":
-                return { ...item, url: `/modeler/#/${workspaceId}` }; // TODO - deep links
-            case "attribute":
-                return { ...item, url: `/modeler/#/${workspaceId}` }; // TODO - deep links
-            case "label":
-                return { ...item, url: `/modeler/#/${workspaceId}` }; // TODO - deep links
-            case "fact":
-                return { ...item, url: `/modeler/#/${workspaceId}` }; // TODO - deep links
-            case "date":
-                return { ...item, url: `/modeler/#/${workspaceId}` }; // TODO - deep links
-            default:
-                return exhaustiveCheck(item.type);
-        }
+        return {
+            ...item,
+            url: getUIPath(item.type, item.id, workspaceId),
+        };
     };
-
-/**
- * Ensure exhaustive switch to prevent missing cases.
- */
-const exhaustiveCheck = (x: never): never => {
-    throw new Error(`Unknown item type ${x}`);
-};
