+++ libs/sdk-ui-semantic-search/src/utils/getUIPath.ts
@@ -1,0 +1,52 @@
+// (C) 2024 GoodData Corporation
+
+/**
+ * Types of objects that we can navigate to in UI.
+ * @internal
+ */
+export type UIPathObjectTypes =
+    | "dashboard"
+    | "visualization"
+    | "metric"
+    | "dataset"
+    | "date"
+    | "fact"
+    | "attribute"
+    | "label";
+
+/**
+ * Get the UI path for the given object type, object ID, and workspace ID.
+ * TODO - this logic should live in gdc-ui repo, would require refactoring of the sdk-ui-semantic-search
+ *  - ui components need URL right away to be able to render native anchor tags
+ *  - gdc-ui would need to inject the URL between the data is loaded and the render
+ * @internal
+ */
+export const getUIPath = (objectType: UIPathObjectTypes, objectId: string, workspaceId: string): string => {
+    switch (objectType) {
+        case "dashboard":
+            return `/dashboards/#/workspace/${workspaceId}/dashboard/${objectId}`;
+        case "visualization":
+            return `/analyze/#/${workspaceId}/${objectId}/edit`;
+        case "metric":
+            return `/metrics/#/${workspaceId}/metric/${objectId}`;
+        case "dataset":
+            return `/modeler/#/${workspaceId}`; // TODO - deep links
+        case "attribute":
+            return `/modeler/#/${workspaceId}`; // TODO - deep links
+        case "label":
+            return `/modeler/#/${workspaceId}`; // TODO - deep links
+        case "fact":
+            return `/modeler/#/${workspaceId}`; // TODO - deep links
+        case "date":
+            return `/modeler/#/${workspaceId}`; // TODO - deep links
+        default:
+            return exhaustiveCheck(objectType);
+    }
+};
+
+/**
+ * Ensure exhaustive switch to prevent missing cases.
+ */
+const exhaustiveCheck = (x: never): never => {
+    throw new Error(`Unknown item type ${x}`);
+};
